+++ Source/ACE/Entity/BroadcastAction.cs
@@ -10,6 +10,8 @@
 
         PlaySound,
 
-        PlayParticleEffect
+        PlayParticleEffect,
+
+        MovementEvent
     }
 }