+++ Source/ACE/Entity/Landblock.cs
@@ -223,6 +223,12 @@ namespace ACE.Entity
             Broadcast(args, true, Quadrant.All);
         }
 
+        public void HandleMovementEvent(WorldObject sender, GeneralMotion motion)
+        {
+            BroadcastEventArgs args = BroadcastEventArgs.CreateMovementEvent(sender, motion);
+            Broadcast(args, true, Quadrant.All);
+        }
+
         public void SendChatMessage(WorldObject sender, ChatMessageArgs chatMessage)
         {
             // only players receive this
@@ -285,6 +291,11 @@ namespace ACE.Entity
                 case BroadcastAction.PlayParticleEffect:
                     {
                         Parallel.ForEach(players, p => p.PlayParticleEffect(args.Effect, args.Sender.Guid));
+                        break;
+                    }
+                case BroadcastAction.MovementEvent:
+                    {
+                        Parallel.ForEach(players, p => p.SendMovementEvent(args.Motion, args.Sender));
                         break;
                     }
             }
@@ -431,6 +442,18 @@ namespace ACE.Entity
                         HandleSoundEvent(obj, soundEffect);
                         break;
                     }
+                case GameActionType.MovementEvent:
+                    {
+                        var g = new ObjectGuid(action.ObjectId);
+                        WorldObject obj = (WorldObject)player;
+                        if (worldObjects.ContainsKey(g))
+                        {
+                            obj = worldObjects[g];
+                        }
+                        var motion = action.Motion;
+                        HandleMovementEvent(obj, motion);
+                        break;
+                    }
                 case GameActionType.QueryHealth:
                     {
                         object target = null;
@@ -504,8 +527,8 @@ namespace ACE.Entity
 
                                             // create the outbound server message
                                             serverMessage = "You have attuned your spirit to this Lifestone. You will resurrect here after you die.";
-                                            player.Session.Network.EnqueueSend(animationEvent, soundEvent); // Slightly doubled sound, why did this get sent in retail?
-                                            // player.Session.Network.EnqueueSend(animationEvent);
+                                            player.EnqueueMovementEvent(motionSanctuary, player.Guid);
+                                            player.Session.Network.EnqueueSend(soundEvent);
                                         }
 
                                         var lifestoneBindMessage = new GameMessageSystemChat(serverMessage, ChatMessageType.Magic);